added some error handling and added test data generation

+++ Payment.cs
@@ -4,8 +4,8 @@ public class Payment
 {
 	public string player1 { get; set; }
 	public string player2 { get; set; }
-	public double amount { get; set; }
-	public Payment(string player1, string player2, double amount)
+	public decimal amount { get; set; }
+	public Payment(string player1, string player2, decimal amount)
 	{
 		this.player1 = player1;
 		this.player2 = player2;
