+++ Player.cs
@@ -3,11 +3,11 @@
 public class Player
 {
     public string name { get; set; }
-    public double buyIn { get; set; }
-    public double chipValue { get; set; }
-    public double balance { get; set; }
+    public decimal buyIn { get; set; }
+    public decimal chipValue { get; set; }
+    public decimal balance { get; set; }
 
-    public Player(string name, double buyIn, double chipValue)
+    public Player(string name, decimal buyIn, decimal chipValue)
     {
         this.name = name;
         this.buyIn = buyIn;
